requirements.txt
Flask
requests
spotipy
pandas
python-dotenv
beautifulsoup4
matplotlib
openai